using bayesflow style and tested build
