+++ tensorflow/contrib/seq2seq/__init__.py
@@ -21,6 +21,6 @@ from __future__ import print_function
 
 import sys
 
-# pylint: disable=unused-import,wildcard-import
-from tensorflow.contrib.seq2seq.python.seq2seq import *
-from tensorflow.contrib.seq2seq.python.kernel_test import *
+# pylint: disable=unused-import,wildcard-import,line-too-long
+from tensorflow.contrib.seq2seq.python.ops import layers
+from tensorflow.contrib.seq2seq.python.ops import loss
