+++ tensorflow/contrib/seq2seq/python/__init__.py
@@ -12,22 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
+"""ops module."""
 
-"""Tests for contrib.seq2seq.python.seq2seq.loss_ops."""
-# pylint: disable=unused-import,g-bad-import-order
 from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
-# pylint: enable=unused-import
-
-import tensorflow as tf
-
-
-class LossTest(tf.case.TestCase):
-
-  def testLoss(self):
-    pass
-
-
-if __name__ == '__main__':
-  tf.test.main()
