--- tensorflow/contrib/seq2seq/python/kernel_test/layers_ops_test.py
@@ -1,36 +1,0 @@
-# Copyright 2016 The TensorFlow Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-
-"""Tests for contrib.seq2seq.python.seq2seq.layers_ops."""
-# pylint: disable=unused-import,g-bad-import-order
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-# pylint: enable=unused-import
-
-import tensorflow as tf
-
-
-class LayersTest(tf.case.TestCase):
-
-  def testRNNDecoder(self):
-    pass
-
-  def testRNNDecoderAttention(self):
-    pass
-
-
-if __name__ == '__main__':
-  tf.test.main()
