+++ tensorflow/contrib/seq2seq/python/ops/layers.py
