+++ tensorflow/contrib/seq2seq/python/ops/loss.py
